Updated Makefile for better testing.
We will use two "test suites". One is a simple hello.c which contains
temporary code for the current development; and the other contains
several source files with code that has been tested before.
The second suite is for regression testing. Now we put the
command to run the regression testing in the Makefile; we should
probably change it to use a separate script (shell or Python) for that.