Updated customized memory malloc&free to automatically handle object ID.
The previous implementation of the customized memory
allocator/deallocator assumes that programmers manually add an ID field
at the beginning of a heap object. However, this is not feasible to
arrays. Now we make the ID of heap objects transparent to programmers:
when a heap object is allocated, we allocate 8 more bytes (for x86-64)
as the ID for it and position the ID right before the real object.
The raw pointer inside the return the mmsafe pointer should still be
pointing to the real beginning of the heap object.