Wrote functions to extract the inner raw pointer inside a mmsafe ptr.
Sometimes we may need to cast a safe pointer to a raw pointer.
If the programmer is completely sure that doing so is completely safe,
these two functions can be used to extrac the inner raw pointer
inside an mmsafe pointer. See the code and comments in the
source file for more details.