Updated mm_alloc/free by the 32-32 key-offset metadata for mm_ptr.
In order to support the situation where using the '&' operator to
get the address of an inner field of a struct, we changed the key
field of an mm_ptr to a metadata filed with both key and offset.
The current implementation adopts a 32-32 key-offset design. But
we plan to support two choices: 32-32 and 40-24.