Fixed a bug of cleaning binaries in olden.sh and olden-origin.sh.
In order to do a new compiling, we need to clean ALL the binaries of
a program. The old code only cleans the final binary of a program
but leaves all the *.o files untouched. This would only trigger
the compiler to do the linking instead of compiling each source file
first. This commit fixed this bug.

Also fixed a path error in olden-origin.sh.