Added mmsafe header files and library to thttpd.
thttpd has a variable "checked" which conflicts with the "checked"
keyword of Checked C. Therefore we keep a separate copy of
the Checked C header file for thttpd and removed the "checked" keyword
in this specific header file. This avoids changing the header file
in the include directory.

Also updated the thttpd_build.sh to link the mmsafe library to the
Makefile of thttpd.