Wrote safe-mm version of realloc() for arrays.
Now we assume that if the re-allocated "new" object starts from the same
starting address of the "old" object, then they are the same object and
hence the "new" object uses the same lock as the "old" one.  If the new
object is really re-allocated and starts from somewhere else, the old
one would be freed and we need create a new lock for the new object.

We may change this assumption to that realloc() always creates a
new object and demands a new key.