Refactored jpegoptim with MMSafe pointers.
Refactored jpegoptim with MMSafe pointers. This program is small
and lexically only calls malloc/realloc 4 times, but one of the
four is in a loop (line 650). We cannot refactor the one in the loop
with our mm-safe version of heap allocator because the return pointer
will be stored in an array that will later be accessed by a libjepg
library function jpeg_read_scanlines(). See this issue for more
details about why this is a problem: https://github.com/jzhou76/checkedc-clang/issues/16

For this program it is fine that we use the original malloc for
the case mentioned above because the allocated data are mainly
accessed by an unchecked library function. The situation gets much
more tricky when some heap object gets touched frequently by both
checked and unchecked code.