[thttpd] Refactored (partial) httpd_get_conn() and httpd_realloc_str().
Refactored httpd_get_conn() to with an mm_ptr argument, and conseqently
refactored (actually wrote a new version of) httpd_realloc_str()
as httpd_get_conn() calls httpd_realloc_str() a lot. The refactoring
of these two functions are not completely done. See the "TODO" comments
in the changed code for more details. In short, we need to refactor
the struct httpd_conn and that seems to be a lot of work as it has
many pointers that need to be changed to mmsafe pointers.