Updated the cmake script of creating llvm-ts build files for CETS.
Added an explicit -O3 to CFLAGS. Now the SoftBoundCETSPass is located
at the end of the Module pass pipeline; when compiling programs with
-O0, this pass would not be run and the libsoftboundcets_rt.so would
have an undefined reference to functions defined by the pass.
The O3.cmake file for Olden only works for Olden; this cmake script
would fail without the explicit -O3 because when it tests if the
compiler works with an small program it uses -O0.