[libsafemm] Fixed a subtle bug in mm_array_realloc().
There was a subtile bug in the old implementation of mm_array_realloc():
it invalidates the old lock after calling realloc() if realloc()
reallocates the memory to a new starting address. This could corrupt
valid memory if realloc has already put new data in the location of
the old lock. This commit fixed this problem by pre-invalidating
the old lock before calling realloc() and then recovering it if
realloc() does not move the starting address of the reallocated memory.