[thttpd] Rewrote the code that converts between checked pointers.
We offered runtime library functions to convert an mm_ptr to an
mm_array_ptr and also the other way around. These are dangerous
operations. For the two cases that use mmptr_to_mmarray_ptr(),
actually we can find other ways to achieve the same goal.
See the code for more details.