[lzfse] Updated how we calculated the overhead for lzfse.
In the last commit (https://github.com/jzhou76/checkedc-safe-mm-misc/commit/53342706de72cd558da958668d8b3ff416476e3f)
we thought we computed the overhead of compression/decompression wrong.
Howefer, on second thoughts, the "corrected" way is actualy not correct.
The overhead of a slower rate b over the original rate a should be
(a - b) /a instead of a / b - 1. For example, the baseline rate is
2 MB/s and the checked rate is 1 MB/s, and intuitively, the rate
slows down by 50% but not 100%.

We also updated how to calculate the geomean. See the changed code
for more details.