[Olden] Revised the scripts for Olden perf and uploaded raw data.
We used one script olden_perf.py to run and collect data for Olden
benchmarks. This is unconvenient because there is no need to
run the baseline benchmarks againa and again. We now have a new script
to run baseline, checked, or cets Olden, and we updated olde_perf.py
to only collect experiments data and calculate results.

Also added raw data files.