[libsafemm] Reverted commit bf04012e0abc0d44458dcf48e2e828e7d5f7f173.
There seems be a compiler bug that caused failing to cast an mmarrayptr
to a raw C pointer when the mmarrayptr is constructed as a rvalue of
pointer arithmetic, e.g., "(int *)(p + i)" will cause an assertion
failure. Specifically, we saw that the reverted commit caused
an assertion failure of the SLPVectorizerPass when optimizing
the call to memmove() in function json_array_remove() of parson
(https://github.com/jzhou76/checkedc-safe-mm-misc/blob/master/benchmarks/checked/parson/parson.c#L1713).