[curl] Added comments on several unported malloc().
Added comments on several malloc() to explain why they are not ported
now. Some are due to unnecessity such as pointer(s) will not escape
before being freed, and some are not ported because they are MIME
related code and we current do not plan to support MIME.
Also there is one case related to dynbuf, which is defined in the lib,
and it is too disruptive to change that now.