[libsafemm] Added a helper function to duplicate a string.
Added a function that takes a raw ptr to char, duplicates the string,
frees the old string, and returns an mm_array_ptr<char> to the new str.
The motivation for having this function is as follows. Sometimes a
library function takes the address of a pointer to a string, allocates
a string on the heap, and makes the ptr arg point to the string.
This would cause errors if we pass the address of an mm_array_ptr to
it because there would be no key-lock metadata generated. One solution
is to write our own mmsafe version of such lib functions. However,
this may be cumbersome because a called lib fn may rely on many other
functions (and global variables, etc.). It could be difficult to
fully port one. The second option is what we did here: simply
duplicating this string and create an mm_array_ptr to it.