[curl] Fixed an environment variable related problem.
If the host system does not have the Nghttp2 library, there would be
two function definition conflicts because we assumed the library
is available and updated the declaration of the two functions
to use checked pointers. This commit keeps both checked and original
function declarations so it would work in systems with and without
the Nghttp2 library.